refactor: consolidate utility functions into AWS Lambda Layer for identity service
