+++ call-profile/main.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 	"time"
 
+	utils "github.com/Real-Dev-Squad/identity-service/layer/utils"
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
-	utils "github.com/rishirishhh/identity-service/layer/utils"
 )
 
 func handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
