+++ scripts/test.sh
@@ -1,9 +1,8 @@
-cd health
 go mod tidy
+cd health
 go test -v
 npx kill-port 8090
 cd ../verify
-go mod tidy
 npm install -g firebase-tools
 if (firebase --project="test" emulators:exec "go test"); then
     echo "Exited Success"
